requirements.txt
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.3
exceptiongroup==1.2.2
google-api-core==2.15.0
google-api-python-client==2.112.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google_spreadsheet==0.0.6
googleapis-common-protos==1.62.0
gspread==5.12.4
h11==0.14.0
httplib2==0.22.0
idna==3.6
influxdb==5.3.1
influxdb-client==1.48.0
msgpack==1.0.7
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
reactivex==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
